requirements/dev.txt
asgiref==3.8.1 ; python_version >= "3.9" and python_version < "4.0"
astroid==2.15.8 ; python_version >= "3.9" and python_version < "4.0"
attrs==24.3.0 ; python_version >= "3.9" and python_version < "4.0"
black==24.10.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation != "PyPy"
cfgv==3.4.0 ; python_version >= "3.9" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0"
coverage==7.6.10 ; python_version >= "3.9" and python_version < "4.0"
cryptography==43.0.3 ; python_version >= "3.9" and python_version < "4.0"
dill==0.3.9 ; python_version >= "3.9" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.9" and python_version < "4.0"
django-cas-sso==1.2.8 ; python_version >= "3.9" and python_version < "4.0"
django-cors-headers==4.6.0 ; python_version >= "3.9" and python_version < "4.0"
django-debug-toolbar==4.4.6 ; python_version >= "3.9" and python_version < "4.0"
django-extensions==3.2.3 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.18 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.9" and python_version < "4.0"
dparse==0.6.4 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.9" and python_version < "4.0"
flake8==7.1.1 ; python_version >= "3.9" and python_version < "4.0"
identify==2.6.5 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
isort==5.13.2 ; python_version >= "3.9" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.9" and python_version < "4.0"
lazy-object-proxy==1.10.0 ; python_version >= "3.9" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.9" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.9" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
pre-commit==3.8.0 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
pycodestyle==2.12.1 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyflakes==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
pylint==2.17.7 ; python_version >= "3.9" and python_version < "4.0"
pyproject-api==1.8.0 ; python_version >= "3.9" and python_version < "4.0"
python-ldap==3.4.4 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.9" and python_version < "4.0"
rpds-py==0.22.3 ; python_version >= "3.9" and python_version < "4.0"
ruamel-yaml-clib==0.2.12 ; platform_python_implementation == "CPython" and python_version < "3.13" and python_version >= "3.9"
ruamel-yaml==0.18.10 ; python_version >= "3.9" and python_version < "4.0"
safety==2.3.4 ; python_version >= "3.9" and python_version < "4.0"
sentry-sdk==2.20.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==75.8.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.2.1 ; python_version >= "3.9" and python_version < "3.11"
tomlkit==0.13.2 ; python_version >= "3.9" and python_version < "4.0"
tox==4.23.2 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2024.2 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
virtualenv==20.29.0 ; python_version >= "3.9" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.9" and python_version < "4.0"
